Added a new render queue category called FPS
The FPS Category has been added to the render queue. This allows materials to be designated as "FPS" in the material settings when using the PSXLit Shader. This will allow any object with an FPS material to render on top of the other ones. This prevents the models from clipping into walls and other surfaces.